A bunch of improvements for iOS 8
1. Only do UITableView layoutSubviews method swizzling on iOS versions before iOS8.
2. Use UITableViewGrouped on iOS 7, due to iOS bug related to section footers.
3. Don't use autolayout for section headers.
4. Improved authentication layout for smaller screens.

+++ ios/ClassNinja/ClassNinja/Classes/CNGenericSelectionViewController.m
@@ -133,7 +133,7 @@
     }
     
     if (VCs.count < searchModels.count) {
-        for (int i = VCs.count; i < searchModels.count; i++) {
+        for (NSUInteger i = VCs.count; i < searchModels.count; i++) {
             cursorVC = [[[cursorVC nextVCClass] alloc] init];
             cursorVC.selectedModel = [searchModels objectAtIndex:i];;
             [self.siongNavigationController pushViewController:cursorVC animated:NO];
